Reorganizing some of the source, adding react-mdl
