+++ package.json
@@ -7,7 +7,8 @@
   },
   "dependencies": {
     "react": "^15.4.0",
-    "react-dom": "^15.4.0"
+    "react-dom": "^15.4.0",
+    "react-mdl": "^1.9.0"
   },
   "scripts": {
     "start": "react-scripts start",
